Split RepositoryValidation into processing actions
