+++ Sources/Application/Kysect.Zeya.Application/LocalHandling/PolicyValidationService.cs
@@ -5,8 +5,9 @@ using Kysect.Zeya.DataAccess.EntityFramework;
 using Kysect.Zeya.GitIntegration.Abstraction;
 using Kysect.Zeya.LocalRepositoryAccess;
 using Kysect.Zeya.RepositoryValidation;
-using Kysect.Zeya.RepositoryValidation.ProcessingActions;
+using Kysect.Zeya.RepositoryValidation.ProcessingActions.CreatePullRequest;
 using Kysect.Zeya.RepositoryValidation.ProcessingActions.Fix;
+using Kysect.Zeya.RepositoryValidation.ProcessingActions.Validation;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
