+++ Sources/Domain/Kysect.Zeya.RepositoryValidation/ProcessingActions/Validation/RepositoryDiagnosticCollector.cs
@@ -1,4 +1,4 @@
-﻿namespace Kysect.Zeya.RepositoryValidation;
+﻿namespace Kysect.Zeya.RepositoryValidation.ProcessingActions.Validation;
 
 public class RepositoryDiagnosticCollector
 {
