+++ Sources/Domain/Kysect.Zeya.RepositoryValidation/ProcessingActions/Validation/RepositoryValidationDiagnostic.cs
@@ -1,3 +1,3 @@
-﻿namespace Kysect.Zeya.RepositoryValidation;
+﻿namespace Kysect.Zeya.RepositoryValidation.ProcessingActions.Validation;
 
 public record RepositoryValidationDiagnostic(string Code, string Repository, string Message, RepositoryValidationSeverity Severity);
