+++ Sources/Domain/Kysect.Zeya.RepositoryValidation/ProcessingActions/Validation/RepositoryValidationReport.cs
@@ -1,7 +1,7 @@
 ﻿using Kysect.CommonLib.BaseTypes.Extensions;
 using System.Diagnostics.Contracts;
 
-namespace Kysect.Zeya.RepositoryValidation;
+namespace Kysect.Zeya.RepositoryValidation.ProcessingActions.Validation;
 
 public record RepositoryValidationReport(IReadOnlyCollection<RepositoryValidationDiagnostic> Diagnostics, IReadOnlyCollection<RepositoryValidationDiagnostic> RuntimeErrors)
 {
