+++ Sources/Domain/Kysect.Zeya.RepositoryValidationRules/Rules/Github/GithubAutoBranchDeletionEnabledValidationRule.cs
@@ -3,6 +3,7 @@ using Kysect.ScenarioLib.Abstractions;
 using Kysect.Zeya.GithubIntegration.Abstraction;
 using Kysect.Zeya.LocalRepositoryAccess.Github;
 using Kysect.Zeya.RepositoryValidation;
+using Kysect.Zeya.RepositoryValidation.ProcessingActions.Validation;
 
 namespace Kysect.Zeya.RepositoryValidationRules.Rules.Github;
