+++ Sources/Domain/Kysect.Zeya.RepositoryValidationRules/Rules/Github/GithubRepositoryLicenseValidationRule.cs
@@ -1,6 +1,7 @@
 ﻿using Kysect.CommonLib.BaseTypes.Extensions;
 using Kysect.ScenarioLib.Abstractions;
 using Kysect.Zeya.RepositoryValidation;
+using Kysect.Zeya.RepositoryValidation.ProcessingActions.Validation;
 
 namespace Kysect.Zeya.RepositoryValidationRules.Rules.Github;
