+++ Sources/Domain/Kysect.Zeya.RepositoryValidationRules/Rules/SourceCode/ArtifactsOutputEnabledValidationRule.cs
@@ -3,6 +3,7 @@ using Kysect.DotnetProjectSystem.SolutionModification;
 using Kysect.ScenarioLib.Abstractions;
 using Kysect.Zeya.LocalRepositoryAccess;
 using Kysect.Zeya.RepositoryValidation;
+using Kysect.Zeya.RepositoryValidation.ProcessingActions.Validation;
 
 namespace Kysect.Zeya.RepositoryValidationRules.Rules.SourceCode;
