+++ Sources/Domain/Kysect.Zeya.RepositoryValidationRules/Rules/SourceCode/NugetVersionSynchronizedWithMasterCentralPackageManagerValidationRule.cs
@@ -5,6 +5,7 @@ using Kysect.DotnetProjectSystem.SolutionModification;
 using Kysect.ScenarioLib.Abstractions;
 using Kysect.Zeya.LocalRepositoryAccess;
 using Kysect.Zeya.RepositoryValidation;
+using Kysect.Zeya.RepositoryValidation.ProcessingActions.Validation;
 using System.ComponentModel.DataAnnotations;
 using System.IO.Abstractions;
