+++ Sources/Tests/Kysect.Zeya.Tests/Application/ValidationPolicyServiceTests.cs
@@ -6,6 +6,7 @@ using Kysect.Zeya.DataAccess.Abstractions;
 using Kysect.Zeya.DataAccess.EntityFramework;
 using Kysect.Zeya.Dtos;
 using Kysect.Zeya.RepositoryValidation;
+using Kysect.Zeya.RepositoryValidation.ProcessingActions.Validation;
 using Kysect.Zeya.Tests.Tools;
 
 namespace Kysect.Zeya.Tests.Application;
