+++ Sources/Tests/Kysect.Zeya.Tests/Domain/RepositoryValidation/PullRequestMessageCreatorTests.cs
@@ -1,6 +1,6 @@
 ﻿using FluentAssertions;
 using Kysect.Zeya.Common;
-using Kysect.Zeya.RepositoryValidation;
+using Kysect.Zeya.RepositoryValidation.ProcessingActions.CreatePullRequest;
 using Kysect.Zeya.RepositoryValidationRules.Rules.SourceCode;
 
 namespace Kysect.Zeya.Tests.Domain.RepositoryValidation;
